Fix hang in Teensy LittleFS initialization (#66)
1. The Teensy build hangs after 'LittleFS: initialized'. I applied the
same fix that's in `fs-pico.cpp` - commenting out the next statement.
2. I cleaned up the formatting on 'LittleFS: initialized'in both
`fs-pico.cpp` and `fs-teensy.cpp`.

I tested this in the Arduino IDE on both Windows and Linux. I don't have
a Mac though. I've also tested it on the command line in Linux, but I'll
probably stick with the IDE going forward.